remove socket.io and add boardgame.io
