+++ src/config.js
@@ -1,0 +1,4 @@
+export const GAME_PORT = process.env.PORT || 8000;
+export const APP_PRODUCTION = false;
+export const GAME_SERVER = `http://localhost:${GAME_PORT}`;
+
