+++ src/pages/Home.js
@@ -85,8 +85,8 @@ function Home({loggedIn,setLoggedIn}) {
             <section className="home subpage">
                 {loggedIn ? (
                     <>
-                    <h3><a href="/hostGame">Host Game</a></h3>
-                    <h3><a href="/joinGame">Join Game</a></h3>
+                    <h3><a href="/play">Play</a></h3>
+                    {/* <h3><a href="/joinGame">Join Game</a></h3> */}
                     <h3><a href="/search">Search For A User</a></h3>
                     <h3><a href="/rules">View Rules</a></h3>
                     </>
